Simplified restart button logic and player objects
There's no need to actually remove the gameboard listener, but it was
interesting to learn about the AbortController and its single-use
AbortSignal object. It's an alternative way to remove and reenable event
listeners. Also, the restart button does not need to clear the player
names, just the board. The player objects have also been simplified to
just name and marker for now until needed.